+++ .vscode/settings.json
@@ -17,5 +17,5 @@
   "files.watcherExclude": {
     "**/.fvm": true
   },
-  "dart.flutterSdkPath": ".fvm/versions/3.22.0"
+  "dart.flutterSdkPath": ".fvm/versions/3.24.3"
 }
