--- lib/src/profile.dart
@@ -1,14 +1,0 @@
-part of '../layrz_models.dart';
-
-@freezed
-class Profile with _$Profile {
-  const factory Profile({
-    String? companyName,
-    String? billingAddress,
-    String? fiscalId,
-    String? headquartersId,
-    Country? headquarters,
-  }) = _Profile;
-
-  factory Profile.fromJson(Map<String, dynamic> json) => _$ProfileFromJson(json);
-}
