+++ lib/src/reports/reports.freezed.dart
@@ -26,12 +26,8 @@ mixin _$ReportPreview {
   /// [pages] is the pages inside of the report
   List<ReportPage> get pages => throw _privateConstructorUsedError;
 
-  /// Serializes this ReportPreview to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-
-  /// Create a copy of ReportPreview
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   $ReportPreviewCopyWith<ReportPreview> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -55,8 +51,6 @@ class _$ReportPreviewCopyWithImpl<$Res, $Val extends ReportPreview>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
-  /// Create a copy of ReportPreview
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -95,8 +89,6 @@ class __$$ReportPreviewImplCopyWithImpl<$Res>
       _$ReportPreviewImpl _value, $Res Function(_$ReportPreviewImpl) _then)
       : super(_value, _then);
 
-  /// Create a copy of ReportPreview
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -156,14 +148,12 @@ class _$ReportPreviewImpl implements _ReportPreview {
             const DeepCollectionEquality().equals(other._pages, _pages));
   }
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   int get hashCode => Object.hash(
       runtimeType, name, const DeepCollectionEquality().hash(_pages));
 
-  /// Create a copy of ReportPreview
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
   _$$ReportPreviewImplCopyWith<_$ReportPreviewImpl> get copyWith =>
@@ -185,18 +175,16 @@ abstract class _ReportPreview implements ReportPreview {
   factory _ReportPreview.fromJson(Map<String, dynamic> json) =
       _$ReportPreviewImpl.fromJson;
 
-  /// [name] is the name of the report preview.
   @override
+
+  /// [name] is the name of the report preview.
   String get name;
+  @override
 
   /// [pages] is the pages inside of the report
-  @override
   List<ReportPage> get pages;
-
-  /// Create a copy of ReportPreview
-  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   _$$ReportPreviewImplCopyWith<_$ReportPreviewImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -216,12 +204,8 @@ mixin _$ReportPage {
   /// [headers] is the headers of the report
   List<ReportHeader> get headers => throw _privateConstructorUsedError;
 
-  /// Serializes this ReportPage to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-
-  /// Create a copy of ReportPage
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   $ReportPageCopyWith<ReportPage> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -245,8 +229,6 @@ class _$ReportPageCopyWithImpl<$Res, $Val extends ReportPage>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
-  /// Create a copy of ReportPage
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -290,8 +272,6 @@ class __$$ReportPageImplCopyWithImpl<$Res>
       _$ReportPageImpl _value, $Res Function(_$ReportPageImpl) _then)
       : super(_value, _then);
 
-  /// Create a copy of ReportPage
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -372,7 +352,7 @@ class _$ReportPageImpl implements _ReportPage {
             const DeepCollectionEquality().equals(other._headers, _headers));
   }
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   int get hashCode => Object.hash(
       runtimeType,
@@ -380,9 +360,7 @@ class _$ReportPageImpl implements _ReportPage {
       const DeepCollectionEquality().hash(_rows),
       const DeepCollectionEquality().hash(_headers));
 
-  /// Create a copy of ReportPage
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
   _$$ReportPageImplCopyWith<_$ReportPageImpl> get copyWith =>
@@ -405,22 +383,20 @@ abstract class _ReportPage implements ReportPage {
   factory _ReportPage.fromJson(Map<String, dynamic> json) =
       _$ReportPageImpl.fromJson;
 
-  /// [name] is the name of the report page.
   @override
+
+  /// [name] is the name of the report page.
   String get name;
+  @override
 
   /// [rows] is the rows inside of the report
-  @override
   List<ReportRow> get rows;
+  @override
 
   /// [headers] is the headers of the report
-  @override
   List<ReportHeader> get headers;
-
-  /// Create a copy of ReportPage
-  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   _$$ReportPageImplCopyWith<_$ReportPageImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -437,12 +413,8 @@ mixin _$ReportRow {
   /// [compact] indicates if the row is compact.
   bool get compact => throw _privateConstructorUsedError;
 
-  /// Serializes this ReportRow to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-
-  /// Create a copy of ReportRow
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   $ReportRowCopyWith<ReportRow> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -465,8 +437,6 @@ class _$ReportRowCopyWithImpl<$Res, $Val extends ReportRow>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
-  /// Create a copy of ReportRow
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -505,8 +475,6 @@ class __$$ReportRowImplCopyWithImpl<$Res>
       _$ReportRowImpl _value, $Res Function(_$ReportRowImpl) _then)
       : super(_value, _then);
 
-  /// Create a copy of ReportRow
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -567,14 +535,12 @@ class _$ReportRowImpl implements _ReportRow {
             (identical(other.compact, compact) || other.compact == compact));
   }
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   int get hashCode => Object.hash(
       runtimeType, const DeepCollectionEquality().hash(_content), compact);
 
-  /// Create a copy of ReportRow
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
   _$$ReportRowImplCopyWith<_$ReportRowImpl> get copyWith =>
@@ -595,18 +561,16 @@ abstract class _ReportRow implements ReportRow {
   factory _ReportRow.fromJson(Map<String, dynamic> json) =
       _$ReportRowImpl.fromJson;
 
-  /// [content] is the cells of the report row.
   @override
+
+  /// [content] is the cells of the report row.
   List<ReportCell> get content;
+  @override
 
   /// [compact] indicates if the row is compact.
-  @override
   bool get compact;
-
-  /// Create a copy of ReportRow
-  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   _$$ReportRowImplCopyWith<_$ReportRowImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -628,12 +592,8 @@ mixin _$ReportHeader {
   @ColorOrNullConverter()
   Color? get color => throw _privateConstructorUsedError;
 
-  /// Serializes this ReportHeader to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-
-  /// Create a copy of ReportHeader
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   $ReportHeaderCopyWith<ReportHeader> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -660,8 +620,6 @@ class _$ReportHeaderCopyWithImpl<$Res, $Val extends ReportHeader>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
-  /// Create a copy of ReportHeader
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -708,8 +666,6 @@ class __$$ReportHeaderImplCopyWithImpl<$Res>
       _$ReportHeaderImpl _value, $Res Function(_$ReportHeaderImpl) _then)
       : super(_value, _then);
 
-  /// Create a copy of ReportHeader
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -775,13 +731,11 @@ class _$ReportHeaderImpl implements _ReportHeader {
             (identical(other.color, color) || other.color == color));
   }
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   int get hashCode => Object.hash(runtimeType, content, textColor, color);
 
-  /// Create a copy of ReportHeader
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
   _$$ReportHeaderImplCopyWith<_$ReportHeaderImpl> get copyWith =>
@@ -804,24 +758,22 @@ abstract class _ReportHeader implements ReportHeader {
   factory _ReportHeader.fromJson(Map<String, dynamic> json) =
       _$ReportHeaderImpl.fromJson;
 
-  /// [content] is the content of the report header.
   @override
+
+  /// [content] is the content of the report header.
   String get content;
+  @override
 
   /// [textColor] is the text color of the report header.
-  @override
   @ColorOrNullConverter()
   Color? get textColor;
+  @override
 
   /// [color] is the color of the report header.
-  @override
   @ColorOrNullConverter()
   Color? get color;
-
-  /// Create a copy of ReportHeader
-  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   _$$ReportHeaderImplCopyWith<_$ReportHeaderImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -850,12 +802,8 @@ mixin _$ReportCell {
   /// [currencySymbol] is the currency symbol of the report cell.
   String? get currencySymbol => throw _privateConstructorUsedError;
 
-  /// Serializes this ReportCell to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-
-  /// Create a copy of ReportCell
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   $ReportCellCopyWith<ReportCell> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -884,8 +832,6 @@ class _$ReportCellCopyWithImpl<$Res, $Val extends ReportCell>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
-  /// Create a copy of ReportCell
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -944,8 +890,6 @@ class __$$ReportCellImplCopyWithImpl<$Res>
       _$ReportCellImpl _value, $Res Function(_$ReportCellImpl) _then)
       : super(_value, _then);
 
-  /// Create a copy of ReportCell
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -1036,7 +980,7 @@ class _$ReportCellImpl implements _ReportCell {
                 other.currencySymbol == currencySymbol));
   }
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   int get hashCode => Object.hash(
       runtimeType,
@@ -1046,9 +990,7 @@ class _$ReportCellImpl implements _ReportCell {
       dataType,
       currencySymbol);
 
-  /// Create a copy of ReportCell
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
   _$$ReportCellImplCopyWith<_$ReportCellImpl> get copyWith =>
@@ -1073,33 +1015,31 @@ abstract class _ReportCell implements ReportCell {
   factory _ReportCell.fromJson(Map<String, dynamic> json) =
       _$ReportCellImpl.fromJson;
 
-  /// [content] is the content of the report header.
   @override
+
+  /// [content] is the content of the report header.
   dynamic get content;
+  @override
 
   /// [textColor] is the text color of the report header.
-  @override
   @ColorOrNullConverter()
   Color? get textColor;
+  @override
 
   /// [color] is the color of the report header.
-  @override
   @ColorOrNullConverter()
   Color? get color;
+  @override
 
   /// [dataType] is the data type of the report cell.
-  @override
   @ReportDataTypeOrNullConverter()
   ReportDataType? get dataType;
+  @override
 
   /// [currencySymbol] is the currency symbol of the report cell.
-  @override
   String? get currencySymbol;
-
-  /// Create a copy of ReportCell
-  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   _$$ReportCellImplCopyWith<_$ReportCellImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
