+++ lib/src/simulation/simulation.freezed.dart
@@ -32,12 +32,8 @@ mixin _$SimulationCycle {
   /// [fileUri] is the URI of the file that contains the cycle
   String? get fileUri => throw _privateConstructorUsedError;
 
-  /// Serializes this SimulationCycle to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-
-  /// Create a copy of SimulationCycle
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   $SimulationCycleCopyWith<SimulationCycle> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -61,8 +57,6 @@ class _$SimulationCycleCopyWithImpl<$Res, $Val extends SimulationCycle>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
-  /// Create a copy of SimulationCycle
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -111,8 +105,6 @@ class __$$SimulationCycleImplCopyWithImpl<$Res>
       _$SimulationCycleImpl _value, $Res Function(_$SimulationCycleImpl) _then)
       : super(_value, _then);
 
-  /// Create a copy of SimulationCycle
-  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -186,13 +178,11 @@ class _$SimulationCycleImpl implements _SimulationCycle {
             (identical(other.fileUri, fileUri) || other.fileUri == fileUri));
   }
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   int get hashCode => Object.hash(runtimeType, id, name, steps, fileUri);
 
-  /// Create a copy of SimulationCycle
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
   _$$SimulationCycleImplCopyWith<_$SimulationCycleImpl> get copyWith =>
@@ -217,26 +207,24 @@ abstract class _SimulationCycle implements SimulationCycle {
   factory _SimulationCycle.fromJson(Map<String, dynamic> json) =
       _$SimulationCycleImpl.fromJson;
 
-  /// [id] is the unique identifier of the cycle
   @override
+
+  /// [id] is the unique identifier of the cycle
   String get id;
+  @override
 
   /// [name] is the name of the cycle
-  @override
   String get name;
+  @override
 
   /// [steps] is the number of steps of the cycle
-  @override
   int get steps;
+  @override
 
   /// [fileUri] is the URI of the file that contains the cycle
-  @override
   String? get fileUri;
-
-  /// Create a copy of SimulationCycle
-  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
+  @JsonKey(ignore: true)
   _$$SimulationCycleImplCopyWith<_$SimulationCycleImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
