+++ lib/src/users/src/user.dart
@@ -1,4 +1,4 @@
-part of '../layrz_models.dart';
+part of '../users.dart';
 
 @freezed
 class User with _$User {
@@ -48,9 +48,6 @@ class User with _$User {
     /// [platformAuth] represents the platform auth.
     @Default("") String platformAuth,
 
-    /// [profile] represents the user profile.
-    Profile? profile,
-
     /// [childs] represents the list of childs.
     @Default([]) List<User> childs,
 
@@ -62,9 +59,6 @@ class User with _$User {
 
     /// [planId] represents the plan ID.
     String? planId,
-
-    /// [configuration] represents the configuration.
-    @EnvironmentOrNullConverter() Environment? configuration,
 
     /// [allowedApps] represents the list of allowed apps.
     List<RegisteredApp>? allowedApps,
@@ -132,82 +126,25 @@ class User with _$User {
   factory User.fromJson(Map<String, dynamic> json) => _$UserFromJson(json);
 }
 
-enum Environment {
-  development,
-  production;
+@unfreezed
+class ProfileInput with _$ProfileInput {
+  factory ProfileInput({
+    @Default('') String name,
+    @Default('') String email,
+    @Default('') String username,
+    AvatarInput? dynamicAvatar,
+  }) = _ProfileInput;
 
-  @override
-  String toString() => toJson();
-
-  String toJson() {
-    switch (this) {
-      case Environment.development:
-        return "DEVELOPMENT";
-      case Environment.production:
-        return "PRODUCTION";
-    }
-  }
-
-  static Environment fromJson(String json) {
-    switch (json) {
-      case "DEVELOPMENT":
-        return Environment.development;
-      case "PRODUCTION":
-        return Environment.production;
-      default:
-        return Environment.development;
-    }
-  }
-}
-
-class EnvironmentOrNullConverter implements JsonConverter<Environment?, String?> {
-  const EnvironmentOrNullConverter();
-
-  @override
-  Environment? fromJson(String? json) {
-    if (json == null) return null;
-    return Environment.fromJson(json);
-  }
-
-  @override
-  String? toJson(Environment? object) => object?.toJson();
-}
-
-class EnvironmentConverter implements JsonConverter<Environment, String> {
-  const EnvironmentConverter();
-
-  @override
-  Environment fromJson(String json) => Environment.fromJson(json);
-
-  @override
-  String toJson(Environment object) => object.toJson();
-}
-
-@freezed
-class InviteLink with _$InviteLink {
-  const InviteLink._();
-  const factory InviteLink({
-    required String id,
-    required String code,
-    String? sentTo,
-    @TimestampOrNullConverter() DateTime? sentAt,
-    @TimestampConverter() required DateTime expiresAt,
-    @TimestampConverter() required DateTime createdAt,
-  }) = _InviteLink;
-
-  factory InviteLink.fromJson(Map<String, dynamic> json) => _$InviteLinkFromJson(json);
-
-  String get url => "https://invite.layrz.com/$code";
+  factory ProfileInput.fromJson(Map<String, dynamic> json) => _$ProfileInputFromJson(json);
 }
 
 @unfreezed
-class TenvioDestinationSetup with _$TenvioDestinationSetup {
-  factory TenvioDestinationSetup({
-    @Default('') String address,
-    double? latitude,
-    double? longitude,
-    @Default(false) bool addressIsPlusCode,
-  }) = _TenvioDestinationSetup;
+class PasswordInput with _$PasswordInput {
+  factory PasswordInput({
+    @Default('') String currentPassword,
+    @Default('') String newPassword,
+    @Default('') String newPasswordConfirmation,
+  }) = _PasswordInput;
 
-  factory TenvioDestinationSetup.fromJson(Map<String, dynamic> json) => _$TenvioDestinationSetupFromJson(json);
+  factory PasswordInput.fromJson(Map<String, dynamic> json) => _$PasswordInputFromJson(json);
 }
